Remove deprecation on standard encoding/decoding
Developer experience matters. Functions for encoding/decoding should be
easily accessible and easily found in docs. With traits it's not
straightforward to figure out how to even use the crate. This commit
doesn't fix a problem of documentation or discoverability of functions
to decode/encode, but for now it at least makes the standard
encode/decode work without fuss.